refactor: remove indicateConditionValues related codes (#281)
* fix(data-cell): cyccle dependence bugs (#269)

* feat: add pointer cursor for link field text (#270)

* chore(ci): ci enhance (#268)

* chore(release): 🔖 publish new version v0.1.5 (#267)

* chore: :fire: remoce useless annotations

* chore: change the build script

* chore(release): :bookmark: publish new version

 - @antv/s2@0.1.5

* chore(ci): pr test

* chore(ci): assign

* chore(ci): add issue ci

* chore(ci): rename

* chore(ci): badge

* chore(ci): auto add reviwer

* chore(ci): update lib link

* chore(ci): fix lint error

* chore: :wrench: update the site preview configuration

* chore: :wrench: add preview workflows

* chore: :wrench: update the site scripts

* chore: :memo: update the tips info for the issue close and issue labeled

* chore(ci): fix bundle size

* chore(ci): update compressed size action

* chore(ci): fix eslint ignore config

* chore(ci): update action

* chore(ci): auto add weekly label

* chore(ci): update node version


* fix: replace circular deps (#271)

* fix: correct indicate condition update logic

* fix: fix import order

* refactor: remove indicateConditionValues

* fix: remove import

* fix: correct import path

* test: fix unit test
