+++ .github/workflows/pr-auto-assign-reviewer.yml
@@ -22,4 +22,4 @@ jobs:
           token: ${{ secrets.GITHUB_TOKEN }}
           emoji: '+1'
           reviewers: 'xingwanying,qubaomingg,yangzhanmei,YardWill,lijinke666,wjgogogo,zxc0328'
-          excludeReviewers: ${{ github.event.pull_request.user.login }}
+          review-creator: false
