+++ packages/s2-core/__tests__/unit/utils/condition/conditions-spec.ts
@@ -1,0 +1,22 @@
+import { getIconPositionCfg } from '@/utils/condition/condition';
+
+describe('getIconLayoutPosition Test', () => {
+  test('should return right by default', () => {
+    expect(
+      getIconPositionCfg({
+        field: 'value',
+        mapping: () => ({ fill: 'red' }),
+      }),
+    ).toEqual('right');
+  });
+
+  test(`should return left when it's left`, () => {
+    expect(
+      getIconPositionCfg({
+        field: 'value',
+        position: 'left',
+        mapping: () => ({ fill: 'red' }),
+      }),
+    ).toEqual('left');
+  });
+});
