+++ packages/s2-core/__tests__/unit/utils/merge.spec.ts
@@ -1,0 +1,8 @@
+import { customMerge } from '@/utils/merge';
+describe('merge test', () => {
+  test('should replace old array with new one', () => {
+    expect(customMerge({ arr: [1, 2, 3] }, { arr: [4, 5, 6] })).toEqual({
+      arr: [4, 5, 6],
+    });
+  });
+});
