+++ packages/s2-core/src/utils/condition/condition.ts
@@ -1,0 +1,5 @@
+import { IconCondition } from '@/common/interface';
+
+export const getIconPositionCfg = (condition: IconCondition) => {
+  return condition?.position ?? 'right';
+};
